Add files via upload
Fires have had a catastrophic effect on both property and human lives for a very long time. As a result, humans have
devised a variety of fire-fighting strategies to address the problem. The novel method forecasts the likelihood of a fire
breaking out. However, predicting the incidence of fire remains a difficult challenge. It's because fire is linked to a
plethora of factors that influence its incidence, and the overall number of fire occurrences every day is quite low. This
indicates that the data set for fire incidence is sparser, as the majority of the data sets' values are zero. As a result, assessing
the risks of fire has been adopted in forecasting the occurrence of fire in order to battle this issue. The machine learning
algorithms used in this approach include boosting, super vector machines, and logistic regression, among others. Fire
detection has attracted significant attention globally because to the health and environmental implications of fires, as well
as the fact that fire regimes have evolved (Castelli et al., 2015). When compared to the existing models for predicting the
incidence of fire, the accuracy of predicting using machine learning models is considerably greater.
Public safety is a major domain which needs constant attention and quick resolutions during various situation arising
at locations such as hospitals, universities, shopping malls and residential areas. Among the various emergencies occurring
on a daily basis, fire prevention is most critical situations and readiness and urgencies by the fire department is of upmost
importance. Considering Boston as an area of study, we see a lot of curiosity in the public minds when there is a fire truck
driving around the block making loud siren and attending the fire alarm. But there is more to this than what meets the eye.
The readiness and urgency from the department’s perspective is a completely different analytical scenario. Be it a true alarm
or a false alarm there is movement of resources under multiple circumstances and there is room for improvement based on
the patterns repeating on a daily basis.